# Improve hash code of Names