The old approach of using the first, last, and middle characters
only lays a trap for generate names that have little or no entropy
at these locations. Fresh existential names generated in "as seen
from" operations are one such case, and when compiling large
batches of files the name table can become imbalanced.

I found such an example in ScalaTest:

   https://github.com/scala/scala-dev/issues/246#issuecomment-255338925

This commit uses all characters to compute the hashCode.

Review by @adriaanm @lrytz

I expect that this change will be beneficial on large code bases
that regularly exercise `AsSeenFrom#captureThis`, and neutral on
most smaller code bases. I'll perform some benchmarks to make sure
of that "neutral" claim and post the results here.