@retronym 2016-10-21
Related to https://github.com/scala/scala-dev/issues/246