@lrytz 2016-10-21
great numbers for ScalaTest, hope there's no negative impact on other projects