@retronym 2016-10-22
My first attempt to benchmark this for small programs showed no change (795ms->797ms) for `better-files`, and a 2% slowdown (1645ms -> 1678ms) for `scalap`. However, the results seem a little noiser on my laptop (even with TurboBoost disabled) than i was used to on my desktop where I've done previous benchmarking, so I'm not sure I have enough warmup time or forks to say whether or not that 2% is just noise.

I plan to re-run the benchmarks, and also compare profiles before/after to see is `hashValue` is starting to show up.