@adriaanm 2016-12-19
Have you had a chance to confirm the benchmark results?