@retronym 2019-04-10
Benchmark run: https://scala-ci.typesafe.com/view/scala-bench/job/compiler-benchmark/2573/console