@retronym 2019-05-08
I'm happy with the performance, and will merge this now.