@lrytz 2019-05-08
> I'm happy with the performance

What did you observe?