@retronym 2019-05-08
@lrytz [neutral to ~1% regression](https://scala-ci.typesafe.com/grafana/dashboard/db/pr-validation?orgId=1&from=1549080972797&to=1555585465453&var-scalaSha=2d2b8959c3%7C3b36ff82c1%7Cc5877a73e7%7C652f49bbfb%7Cc2d25ac469%7Cf6d09146bd%7Cebef3cddc8%7Cf0ec40c1cc%7Cfff4dc0391%7C82e6d0871a%7C51230ecbeb%7C48c3e1dba1&var-source=All&var-bench=HotScalacBenchmark.compile&var-host=scalabench%40scalabench%40)

I'm hopeful that #8019 will recover a little of the regression (if it is real, that is!). I'll keep an eye on the performance charts.