@retronym 2019-05-08
The motivation to revive this, BTW, was that we had a report from a customer that they had a name table bucket with 500 entries.